refactor: move `deps` test case
